feat: add mutate existing API
