+++ api/policies.kyverno.io/v1alpha1/validating_policy.go
@@ -5,13 +5,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-type EvaluationMode string
-
-const (
-	EvaluationModeKubernetes EvaluationMode = "Kubernetes"
-	EvaluationModeJSON       EvaluationMode = "JSON"
-)
-
 // +genclient
 // +genclient:nonNamespaced
 // +kubebuilder:object:root=true
@@ -258,37 +251,4 @@ type WebhookConfiguration struct {
 	// After the configured time expires, the admission request may fail, or may simply ignore the policy results,
 	// based on the failure policy. The default timeout is 10s, the value must be between 1 and 30 seconds.
 	TimeoutSeconds *int32 `json:"timeoutSeconds,omitempty"`
-}
-
-type EvaluationConfiguration struct {
-	// Mode is the mode of policy evaluation.
-	// Allowed values are "Kubernetes" or "JSON".
-	// Optional. Default value is "Kubernetes".
-	// +optional
-	Mode EvaluationMode `json:"mode,omitempty"`
-
-	// Admission controls policy evaluation during admission.
-	// +optional
-	Admission *AdmissionConfiguration `json:"admission,omitempty"`
-
-	// Background  controls policy evaluation during background scan.
-	// +optional
-	Background *BackgroundConfiguration `json:"background,omitempty"`
-}
-
-type AdmissionConfiguration struct {
-	// Enabled controls if rules are applied during admission.
-	// Optional. Default value is "true".
-	// +optional
-	// +kubebuilder:default=true
-	Enabled *bool `json:"enabled,omitempty"`
-}
-
-type BackgroundConfiguration struct {
-	// Enabled controls if rules are applied to existing resources during a background scan.
-	// Optional. Default value is "true". The value must be set to "false" if the policy rule
-	// uses variables that are only available in the admission review request (e.g. user name).
-	// +optional
-	// +kubebuilder:default=true
-	Enabled *bool `json:"enabled,omitempty"`
 }
